# Chapter 15 part 1: OBJ parser for regular triangles and KD Trees for Groups representing triangular meshes