--- a/src/main/kotlin/apps/ch14_hexagon.kt
+++ b/src/main/kotlin/apps/ch14_hexagon.kt
@@ -17,7 +17,7 @@ fun main() {
         val sides = (0 until 6).map { Matrix.rotateY(it * PI / 3) }.map {
             side.withTransformation(it)
         }
-        Group(Matrix.rotateX(-0.4363) * Matrix.rotateY(-PI / 18), children = sides)
+        Group(sides,Matrix.rotateX(-0.4363) * Matrix.rotateY(-PI / 18))
     }
 
     val world = run {