--- a/src/test/kotlin/shapes/TriangleTest.kt
+++ b/src/test/kotlin/shapes/TriangleTest.kt
@@ -1,0 +1,4 @@
+package shapes
+
+class TriangleTest {
+}