… output/ch14_hexagon.png
… output/ch15_world1.png
… src/main/resources/teddybear.obj
… src/test/kotlin/input/OBJParserTest.kt
… src/test/kotlin/shapes/GroupTest.kt
… src/test/kotlin/shapes/ShapeTest.kt
… src/test/resources/faces.obj
… src/test/resources/gibberish.obj
… src/test/resources/points.obj
… src/test/resources/polygon.obj
… src/test/resources/triangles.obj