- Cleaned up part of Chapter 14, namely added hexagon image and refactored the two examples to share common code creating `leg` aka `side`.
- Rearranged Group parameters to have `children` first and a possible `transformation` second.
- Added `Triangle`, `TriangleTest`, `OBJParser`, and test cases to render triangle meshes.
- Added primitive `KDTree` for `Group`s that have only `Triangle`s as children to reduce computation time.

- `KDTree` use massively improves performance for `ch15_world1` rendering:
- For resolution 400x400, rendering time improves from 69s to 3.5s (95% reduction).
- For resolution 1200x1200, rendering time improves from 563s to 23s (96% reduction).